Started properly implementing addActivity
This is like the 3rd time I'm back to this, because I either had trouble with Cucumbeer duplicate steps, with merging, or with not merging into Main, giving me more Cucumber trouble.

The files have been added and the non-activity tests thankfully run without issue, but my own tests are dying again.
I believe it's because of DateServer not getting its "calender"-variable instantiated before it gets called through the Project-class upon project crreation for the tests - I'm getting tons of NullPointers, all seeming to trail back to getYer() in DateServer.