Changed String comparison
This has bugged me for too long, so I refactored the "==" between two String types in, and removed a pointless method.

I know this is not relevant to deleting Activities, but I refactored the code, so it still runs without issues before and after.
Changes aren't very significant so they're unlikely to cause issues for others.